+++ zsh/install.sh
@@ -1,26 +1,62 @@
 #!/bin/bash
 set -e
-if [[ -z "${DOTFILES}" ]]; then
-    export DOTFILES="$(git rev-parse --show-toplevel)"
-fi
-source "$DOTFILES/.lib/include.sh"
+for file in $(git rev-parse --show-toplevel)/lib/{log,configurations}.sh; do
+    source "${file}"
+done
 
-function _install_powerline_font { (
-    cd /tmp || return
-    git clone https://github.com/powerline/fonts.git --depth=1
-    cd fonts
-    ./install.sh
-    cd ..
-    rm -rf fonts
+function _get_font_dir {
+    if [ ! "$(uname)" = "Darwin" ]; then
+            mkdir -p ~/.local/share/fonts
+            dir="$HOME/.local/share/fonts"
+    else
+            dir="$HOME/Library/Fonts"
+    fi
+    echo $dir
+}
+
+function _install_powerline_fonts { (
+    if [ ! -f "$(_get_font_dir)/Hack-Regular.ttf" ]; then
+        _log_info "Installing powerline fonts"
+        cd /tmp
+        git clone https://github.com/powerline/fonts.git --depth=1
+        cd fonts
+        ./install.sh
+        cd ..
+        rm -rf fonts
+    fi
 ) }
 
-function main {
-    if ! zsh --version &> /dev/null; then
-        install_package_universal zsh
-        chsh -s /bin/zsh $USER #set zsh as current default shell
+function _install_nerd_fonts { (
+    font_dir="$(_get_font_dir)"
+    if [ ! -f "$font_dir/Droid Sans Mono for Powerline Nerd Font Complete.otf" ]; then
+        _log_info "Installing nerd fonts"
+        cd $font_dir
+        curl -fLo "Droid Sans Mono for Powerline Nerd Font Complete.otf" https://github.com/ryanoasis/nerd-fonts/raw/master/patched-fonts/DroidSansMono/complete/Droid%20Sans%20Mono%20Nerd%20Font%20Complete.otf
     fi
-    _install_powerline_fonts
+) }
+
+function _install_brew {
+    if ! which brew &> /dev/null; then
+        _log_info "Installing brew"
+        if [ "$(uname)" = "Linux" ]; then
+            sh -c "$(curl -fsSL https://raw.githubusercontent.com/Linuxbrew/install/master/install.sh)" # Install brew if not installed
+        elif [ "$(uname)" = "Darwin" ]; then
+            ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
+        fi
+    else
+    fi
+}
+
+function main {
+    _install_brew
+    if ! which zsh &> /dev/null; then
+        brew install zsh
+    fi
     link_config ".zshrc"
+    _log_info "setting zsh as a default shell"
+    chsh -s /bin/zsh $USER
+    _install_powerline_fonts
+    _install_nerd_fonts
 }
 
 main
